+++ src/internal_types.rs
@@ -1,0 +1,4 @@
+pub struct TableWrapper {
+    pub path_string: String,
+    pub table: prettytable::Table,
+}